ModelAI/Model2.0/requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.14
tensorflow==2.17.0
numpy==1.26.4
scikit-learn==1.4.2
pyttsx3==2.90
tk==0.1.0
pytube==15.0.0
pytubefix==7.4.0
Pillow==10.3.0